Move Protected and Unauthenticatedlayout into a directory
