+++ src/router.tsx
@@ -1,6 +1,6 @@
 import {createBrowserRouter} from "react-router-dom";
-import ProtectedLayout from "./Components/ProtectedLayout";
-import UnauthenticatedLayout from "./Components/UnauthenticatedLayout";
+import ProtectedLayout from "./Components/Layout/ProtectedLayout";
+import UnauthenticatedLayout from "./Components/Layout/UnauthenticatedLayout";
 import Login from "./screens/Login/Login";
 import TaskScreen from "./screens/Tasks/TaskScreen";
